Feature #70119, VariableAssemblyDialogController service proxy (#630)
