+++ core/src/main/java/inetsoft/web/composer/model/vs/VariableListDialogModel.java
@@ -19,8 +19,10 @@ package inetsoft.web.composer.model.vs;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.io.Serializable;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class VariableListDialogModel {
+public class VariableListDialogModel implements Serializable {
    public String[] getLabels() {
       if(labels == null) {
          labels = new String[0];
